Move BrowserStandardEncoder
