+++ src/frontend/packages/core/src/features/endpoints/backup-restore/backup-endpoints.service.ts
@@ -5,9 +5,9 @@ import { BehaviorSubject, Observable } from 'rxjs';
 import { first, map } from 'rxjs/operators';
 
 import { GeneralEntityAppState } from '../../../../../store/src/app-state';
+import { BrowserStandardEncoder } from '../../../../../store/src/browser-encoder';
 import { entityCatalog } from '../../../../../store/src/entity-catalog/entity-catalog';
 import { EndpointModel } from '../../../../../store/src/types/endpoint.types';
-import { BrowserStandardEncoder } from '../../../helper';
 import {
   BackupEndpointConfigUI,
   BackupEndpointConnectionTypes,
