+++ src/frontend/packages/core/src/features/endpoints/backup-restore/restore-endpoints.service.ts
@@ -5,10 +5,10 @@ import { BehaviorSubject, combineLatest, Observable } from 'rxjs';
 import { filter, map, switchMap } from 'rxjs/operators';
 
 import { GeneralEntityAppState } from '../../../../../store/src/app-state';
+import { BrowserStandardEncoder } from '../../../../../store/src/browser-encoder';
 import { selectSessionData } from '../../../../../store/src/reducers/auth.reducer';
 import { SessionData } from '../../../../../store/src/types/auth.types';
 import { LoggerService } from '../../../core/logger.service';
-import { BrowserStandardEncoder } from '../../../helper';
 
 interface BackupContent {
   payload: string;
