+++ src/frontend/packages/store/src/browser-encoder.ts
@@ -1,14 +1,5 @@
 import { HttpParameterCodec } from '@angular/common/http';
 
-export function getPath(object: { [key: string]: any }, path: string[]) {
-  try {
-    return path.reduce((o, i) => o[i], object);
-  } catch (err) {
-    return null;
-  }
-}
-
-
 export class BrowserStandardEncoder implements HttpParameterCodec {
   encodeKey(key: string): string {
     return encodeURIComponent(key);
